fix: move DiscountApplicationStrategy to constants
