+++ .changeset/ninety-tables-grin.md
@@ -1,0 +1,5 @@
+---
+'@shopify/discount-app-components': patch
+---
+
+move strategy to constants
